Split webgl into separate crate
