+++ components/webgl/lib.rs
@@ -4,7 +4,10 @@
 
 #![deny(unsafe_code)]
 
-mod raqote_backend;
+pub use webgl_mode::WebGLComm;
 
-pub mod canvas_data;
-pub mod canvas_paint_thread;
+mod webgl_limits;
+mod webgl_mode;
+pub mod webgl_thread;
+#[cfg(feature = "webxr")]
+mod webxr;
